SOFT-233 Четырёхбайтовая адресация для счётчиков Милур
- рефакторинг установки соединения для milur и mercury230

+++ wb-mqtt-serial/em_device.cpp
@@ -6,7 +6,7 @@
 TEMDevice::TEMDevice(PDeviceConfig device_config, PAbstractSerialPort port)
     : TSerialDevice(device_config, port), Config(device_config) {}
 
-void TEMDevice::EnsureSlaveConnected(uint8_t slave, bool force)
+void TEMDevice::EnsureSlaveConnected(uint32_t slave, bool force)
 {
     if (!force && ConnectedSlaves.find(slave) != ConnectedSlaves.end())
         return;
