+++ wb-mqtt-serial/em_device.h
@@ -10,7 +10,8 @@
 #include "serial_device.h"
 
 // Common device base for electricity meters
-class TEMDevice: public TSerialDevice {
+class TEMDevice : public TSerialDevice
+{
 public:
     static const int DefaultTimeoutMs = 1000;
 
@@ -18,12 +19,13 @@ public:
     virtual void WriteRegister(PRegister reg, uint64_t value) override;
 
 protected:
-    enum ErrorType {
+    enum ErrorType
+    {
         NO_ERROR,
         NO_OPEN_SESSION,
         OTHER_ERROR
     };
-    virtual bool ConnectionSetup(uint8_t slave) = 0;
+    virtual bool ConnectionSetup(uint32_t slave) = 0;
     virtual ErrorType CheckForException(uint8_t* frame, int len, const char** message) = 0;
     void WriteCommand(uint8_t slave, uint8_t cmd, uint8_t* payload, int len);
     bool ReadResponse(uint8_t slave, int expectedByte1, uint8_t* buf, int len,
@@ -35,9 +37,9 @@ protected:
     PDeviceConfig DeviceConfig() const;
 
 private:
-    void EnsureSlaveConnected(uint8_t slave, bool force = false);
+    void EnsureSlaveConnected(uint32_t slave, bool force = false);
 
-    std::unordered_set<uint8_t> ConnectedSlaves;
+    std::unordered_set<uint32_t> ConnectedSlaves;
     PDeviceConfig Config;
 };
