+++ wb-mqtt-serial/milur_device.h
@@ -28,7 +28,7 @@ public:
 
 
 protected:
-    virtual bool ConnectionSetup(uint8_t slave) override;
+    virtual bool ConnectionSetup(uint32_t slave) override;
     virtual ErrorType CheckForException(uint8_t* frame, int len, const char** message) override;
     uint64_t BuildIntVal(uint8_t* p, int sz) const;
     uint64_t BuildBCB32(uint8_t* psrc) const;
